backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
motor==3.4.0
pymongo==4.7.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pdfplumber==0.11.0
python-dotenv==1.0.1
anthropic==0.28.0
pydantic[email]==2.7.4
python-docx==1.1.2
PyPDF2==3.0.1
httpx==0.27.0